fix errant tests locally

+++ r2r/main/r2r_app.py
@@ -401,7 +401,8 @@ class R2RApp(metaclass=AsyncSyncMeta):
             self.run_manager, "ingest_documents_app"
         ) as run_id:
             try:
-                return await self.aingest_documents(request.documents)
+                results = await self.aingest_documents(request.documents)
+                return {"results": results}
 
             except HTTPException as he:
                 raise he
@@ -745,7 +746,7 @@ class R2RApp(metaclass=AsyncSyncMeta):
                 )
 
                 # Call aingest_files with the correct order of arguments
-                results = self.aingest_files(
+                results = await self.aingest_files(
                     files=files,
                     metadatas=metadatas,
                     document_ids=ids_list,
@@ -1216,6 +1217,8 @@ class R2RApp(metaclass=AsyncSyncMeta):
         **kwargs: Any,
     ):
         ids = self.providers.vector_db.delete_by_metadata(keys, values)
+        if len(ids) == 0:
+            raise ValueError("No entries found for deletion.")
         self.providers.vector_db.delete_documents_info(ids)
 
         return {"results": "Entries deleted successfully."}
